+++ index.d.ts
@@ -122,14 +122,6 @@ declare function t(strings: any | string[], ...expr: any[]): string;
  */
 declare function l(strings: any | string[], ...expr: any[]): string;
 
-declare class IdleDeadline {
-	readonly didTimeout: boolean;
-	timeRemaining(): number;
-}
-
-declare function requestIdleCallback(fn: (deadline: IdleDeadline) => void, opts?: {timeout?: number}): number;
-declare function cancelIdleCallback(id: number): void;
-
 declare function setImmediate(fn: AnyFunction): number;
 declare function clearImmediate(id: number): void;